LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/at-property-typedom-expected.txt


FAIL Properties declared with @property reify correctly assert_equals: expected " 100px" but got "100px"
FAIL Re-declaring a property with a different type affects reification assert_equals: expected "CSSUnitValue" but got "CSSUnparsedValue"
